Fix event handling in stack example
The help text in the stack example suggests clicking on the windows but
this did not do anything as raw click events were expected while the
terminal was not in raw mouse event mode.

Let the model respond to the SingleClick event instead of the MousePress
so click messages are shown.

Rearrange the print order and clear the background of widgets so that the
borders of the widgets in the lower stack are not shown through and the
end of the previous click message does not remain showing when the next
click is registered.

This combined with listening to resize events also allows the layout to
remain consistent when the terminal is resized.